client/pyproject.toml
[project]
name = "wraith-client"
version = "0.1.0"
description = "WRAI.TH Client — Agent session controller, Discord bridge, and Mission Control UI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "httpx-sse>=0.4.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "pyyaml>=6.0",
    "discord.py>=2.4.0",
    "structlog>=24.0.0",
    "aiosqlite>=0.20.0",
    "websockets>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.6.0",
]

[project.scripts]
wraith-client = "src.main:main"

[tool.ruff]
target-version = "py311"
line-length = 100
